# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agenvas API
  version: 0.1.0
  description: |
    Agenvas 的权威 HTTP 合约。覆盖当前已实现的身份、项目、画布、Agent、Task 与私有素材切片；未实现的 MVP 能力继续按开发清单加入。
    部署者设置 AGENVAS_RECOVERY_MODE=true 后，/api/v1/projects、/api/v1/settings 与 /api/v1/media-templates 下的非 GET/HEAD/OPTIONS 请求统一返回 HTTP 503 ProblemDetail，code=RECOVERY_MODE_READ_ONLY；身份登录与只读核对仍可用。恢复模式同时关闭后台调度，不能把 503 当作业务版本冲突重试。
    服务端为每次 HTTP 请求生成 X-Request-Id 响应头；ProblemDetail.traceId 与它相同。客户端同名请求头不会被信任或回显。
    所有 API 接受 Accept-Language 语言协商（en、zh、ru、ja，地区变体按基础语言匹配；不支持、缺失或无效时回退 zh，q=0 不选用）。响应携带 Content-Language 与 Vary: Accept-Language，包括认证和 CSRF 错误。Problem 的 title、detail、fieldErrors.message 使用协商语言；code、状态、字段路径、ID 和用户内容不随语言变化。语言偏好仅由浏览器保存，无账号配置 API 或数据库迁移。
servers:
  - url: /
paths:
...
  /api/v1/library/entries:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
    get:
      operationId: "listLibraryEntries"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryPage"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "category"
          in: "query"
          schema:
            $ref: "#/components/schemas/LibraryCategory"
        - name: "kind"
          in: "query"
          schema:
            $ref: "#/components/schemas/ArtifactKind"
        - name: "query"
          in: "query"
          schema:
            type: "string"
            maxLength: 160
            default: ""
        - name: "favorite"
          in: "query"
          schema:
            type: "boolean"
            default: false
        - name: "trash"
          in: "query"
          schema:
            type: "boolean"
            default: false
        - name: "mediaOnly"
          in: "query"
          description: "Limit items, total and category counts to IMAGE/VIDEO/AUDIO for asset-page browsing; saved entries and reference pickers are unaffected."
          schema: {type: boolean, default: false}
        - name: "sort"
          in: "query"
          schema:
            $ref: "#/components/schemas/LibrarySort"
            default: "SAVED"
        - name: "cursor"
          in: "query"
          schema:
            type: "string"
            maxLength: 1000
  /api/v1/library/entries/{entryId}:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
    get:
      operationId: "getLibraryEntry"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryEntry"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "entryId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
    patch:
      operationId: "updateLibraryEntry"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryEntry"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "entryId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLibraryRequest"
    delete:
      operationId: "deleteLibraryEntry"
      responses:
        "204":
          description: "Success"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "entryId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
        - name: "expectedVersion"
          in: "query"
          schema:
            type: "integer"
            format: "int64"
            minimum: 0
          required: true
  /api/v1/library/entries/{entryId}/trash:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
    post:
      operationId: "trashLibraryEntry"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryEntry"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "entryId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryVersionRequest"
  /api/v1/library/entries/{entryId}/restore:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
    post:
      operationId: "restoreLibraryEntry"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryEntry"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "entryId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryVersionRequest"
  /api/v1/library/entries/{entryId}/content:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
    get:
      operationId: "getLibraryContent"
      responses:
        "200":
          description: "Private media bytes; content supports Range and HEAD"
          content:
            application/octet-stream:
              schema:
                type: "string"
                format: "binary"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "206":
          description: "Requested byte range"
        "416":
          description: "Invalid byte range"
      parameters:
        - name: "entryId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
    head:
      operationId: "headLibraryContent"
      responses:
        "200":
          description: "Success"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "entryId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
  /api/v1/library/entries/{entryId}/thumbnail:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
    get:
      operationId: "getLibraryThumbnail"
      responses:
        "200":
          description: "Private media bytes; content supports Range and HEAD"
          content:
            application/octet-stream:
              schema:
                type: "string"
                format: "binary"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "entryId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
  /api/v1/library/commands/{commandId}:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
    get:
      operationId: "getLibraryCommand"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryCommand"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "commandId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
  /api/v1/library/commands/{commandId}/retry:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
    post:
      operationId: "retryLibraryCommand"
      responses:
        "202":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryCommand"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "commandId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
  /api/v1/projects/{projectId}/canvas-items/{itemId}/library-saves:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
    get:
      operationId: "getLibrarySaveSource"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibrarySource"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "projectId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
        - name: "itemId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
    post:
      operationId: "saveLibraryEntry"
      responses:
        "202":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryCommand"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "projectId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
        - name: "itemId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveLibraryRequest"
  /api/v1/projects/{projectId}/library-imports:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
    post:
      operationId: "importLibraryEntry"
      responses:
        "202":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryCommand"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "projectId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportLibraryRequest"
  /api/v1/projects/{projectId}/canvas-items/{itemId}/library-references:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
    post:
      operationId: "referenceLibraryEntry"
      responses:
        "202":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryCommand"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      parameters:
        - name: "projectId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
        - name: "itemId"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferenceLibraryRequest"
  /api/v1/library/uploads:
    parameters:
      - {$ref: "#/components/parameters/AcceptLanguage"}
    post:
      operationId: "uploadLibraryEntry"
      responses:
        "202":
          description: "Accepted local transfer. PNG/JPEG/WebP:20 MiB/40 MP; MP3/WAV/OGG Opus:50 MiB/10min; decoded MP4:50 MiB/40 MP/60sec."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryCommand"
        default:
          description: "ProblemDetail error"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: "object"
              additionalProperties: false
              required: ["file", "kind", "name", "category", "commandKey"]
              properties:
                file:
                  type: "string"
                  format: "binary"
                kind:
                  type: "string"
                  enum: ["IMAGE", "VIDEO", "AUDIO"]
                name:
                  type: "string"
                  minLength: 1
                  maxLength: 160
                category:
                  $ref: "#/components/schemas/LibraryCategory"
                commandKey:
                  type: "string"
                  minLength: 1
                  maxLength: 200
...
components:
  schemas:
...
    LibraryCategory:
      type: "string"
      enum: ["CHARACTER", "SCENE", "PROP", "OTHER"]
    LibrarySort:
      type: "string"
      enum: ["SAVED", "NAME", "UPDATED"]
    LibraryEntry:
      type: "object"
      additionalProperties: false
      required: ["id", "name", "category", "kind", "textContent", "contentType", "byteSize", "width", "height", "durationMs", "hasThumbnail", "source", "favorite", "trashedAt", "version", "createdAt", "updatedAt"]
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
          minLength: 1
          maxLength: 160
        category:
          $ref: "#/components/schemas/LibraryCategory"
        kind:
          $ref: "#/components/schemas/ArtifactKind"
        textContent:
          type: ["object", "null"]
          additionalProperties: true
        contentType:
          type: ["string", "null"]
        byteSize:
          type: ["integer", "null"]
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        durationMs:
          type: ["integer", "null"]
        hasThumbnail:
          type: "boolean"
        source:
          type: "object"
          additionalProperties: true
        favorite:
          type: "boolean"
        trashedAt:
          type: ["string", "null"]
          format: "date-time"
        version:
          type: "integer"
          format: "int64"
          minimum: 0
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
    LibraryPage:
      type: "object"
      additionalProperties: false
      required: ["items", "nextCursor", "total", "categoryCounts"]
      properties:
        items:
          type: "array"
          items:
            $ref: "#/components/schemas/LibraryEntry"
        nextCursor:
          type: ["string", "null"]
        total:
          type: "integer"
          minimum: 0
        categoryCounts:
          type: "object"
          additionalProperties:
            type: "integer"
            minimum: 0
    LibraryCommand:
      type: "object"
      additionalProperties: false
      required: ["id", "status", "result", "errorCode", "errorDetail"]
      properties:
        id:
          type: "string"
          format: "uuid"
        status:
          type: "string"
          enum: ["ACCEPTED", "ARCHIVING", "SUCCEEDED", "FAILED"]
        result:
          type: ["object", "null"]
          additionalProperties: false
          required: []
          properties:
            entryId:
              type: "string"
              format: "uuid"
            artifactId:
              type: "string"
              format: "uuid"
            versionId:
              type: "string"
              format: "uuid"
            canvasItemId:
              type: "string"
              format: "uuid"
            draftVersion:
              type: "integer"
              format: "int64"
              minimum: 0
            alreadySaved:
              type: "boolean"
            trashed:
              type: "boolean"
        errorCode:
          type: ["string", "null"]
        errorDetail:
          type: ["string", "null"]
          description: Public failure explanation localized using Accept-Language. Internal message keys and interpolation arguments are never exposed.
    LibrarySource:
      type: "object"
      additionalProperties: false
      required: ["title", "kind", "versionId", "expectedSelectionEpoch", "expectedArtifactVersion", "textContent", "assetId", "versionNo"]
      properties:
        title:
          type: "string"
          minLength: 1
          maxLength: 160
        kind:
          $ref: "#/components/schemas/ArtifactKind"
        versionId:
          type: "string"
          format: "uuid"
        expectedSelectionEpoch:
          type: "integer"
          format: "int64"
          minimum: 0
        expectedArtifactVersion:
          type: "integer"
          format: "int64"
          minimum: 0
        textContent:
          type: ["object", "null"]
          additionalProperties: true
        assetId:
          type: ["string", "null"]
          format: "uuid"
        versionNo:
          type: "integer"
          minimum: 1
    SaveLibraryRequest:
      type: "object"
      additionalProperties: false
      required: ["versionId", "expectedSelectionEpoch", "name", "category", "commandKey"]
      properties:
        versionId:
          type: "string"
          format: "uuid"
        expectedSelectionEpoch:
          type: "integer"
          format: "int64"
          minimum: 0
        expectedArtifactVersion:
          type: "integer"
          format: "int64"
          minimum: 0
        name:
          type: "string"
          minLength: 1
          maxLength: 160
        category:
          $ref: "#/components/schemas/LibraryCategory"
        commandKey:
          type: "string"
          minLength: 1
          maxLength: 200
    UpdateLibraryRequest:
      type: "object"
      additionalProperties: false
      required: ["expectedVersion", "name", "category", "favorite"]
      properties:
        expectedVersion:
          type: "integer"
          format: "int64"
          minimum: 0
        name:
          type: "string"
          minLength: 1
          maxLength: 160
        category:
          $ref: "#/components/schemas/LibraryCategory"
        favorite:
          type: "boolean"
    LibraryVersionRequest:
      type: "object"
      additionalProperties: false
      required: ["expectedVersion"]
      properties:
        expectedVersion:
          type: "integer"
          format: "int64"
          minimum: 0
    ImportLibraryRequest:
      type: "object"
      additionalProperties: false
      required: ["entryId", "expectedVersion", "x", "y", "commandKey"]
      properties:
        entryId:
          type: "string"
          format: "uuid"
        expectedVersion:
          type: "integer"
          format: "int64"
          minimum: 0
        x:
          type: "number"
          minimum: -1000000
          maximum: 1000000
        y:
          type: "number"
          minimum: -1000000
          maximum: 1000000
        commandKey:
          type: "string"
          minLength: 1
          maxLength: 200
    ReferenceLibraryRequest:
      type: "object"
      additionalProperties: false
      required: ["entryId", "expectedVersion", "draft", "role", "color", "commandKey"]
      properties:
        slotKey: {type: string, pattern: "^[A-Za-z][A-Za-z0-9_]{0,63}$", description: "Optional active named workflow media slot; the imported project-local version is assigned atomically with the draft update. Omit for ordinary media references."}
        entryId:
          type: "string"
          format: "uuid"
        expectedVersion:
          type: "integer"
          format: "int64"
          minimum: 0
        draft:
          $ref: "#/components/schemas/SaveMediaDraftRequest"
        role:
          type: "string"
          enum: ["REFERENCE", "START_FRAME", "END_FRAME", "AUDIO_REFERENCE", "VIDEO_REFERENCE"]
        color:
          type: "string"
          pattern: "^#[0-9A-F]{6}$"
        commandKey:
          type: "string"
          minLength: 1
          maxLength: 200
...
    ArtifactKind:
      type: string
      enum: [TEXT, IMAGE, VIDEO, AUDIO]
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, code, traceId, retryable]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        code: {type: string}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        traceId: {type: string}
        retryable: {type: boolean}
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, message]
            properties:
              field: {type: string}
              message: {type: string}
...
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: 支持 en、zh、ru、ja 及其地区变体，按质量权重选择；缺失、不支持或无效时回退 zh。
      schema: {type: string}
      example: ru-RU, en;q=0.8
...